+++ internal/cgen/auxiliary/json.cc
@@ -446,7 +446,7 @@ DecodeJson(DecodeJsonCallbacks& callbacks,
               goto done;
             }
             depth++;
-            if (depth > WUFFS_JSON__DECODER_DEPTH_MAX_INCL) {
+            if (depth > (int32_t)WUFFS_JSON__DECODER_DEPTH_MAX_INCL) {
               ret_error_message =
                   "wuffs_aux::DecodeJson: internal error: bad depth";
               goto done;
