+++ release/c/wuffs-unsupported-snapshot.c
@@ -64623,7 +64623,7 @@ DecodeCbor(DecodeCborCallbacks& callbacks,
               goto done;
             }
             depth++;
-            if (depth > WUFFS_CBOR__DECODER_DEPTH_MAX_INCL) {
+            if (depth > (int32_t)WUFFS_CBOR__DECODER_DEPTH_MAX_INCL) {
               ret_error_message =
                   "wuffs_aux::DecodeCbor: internal error: bad depth";
               goto done;
@@ -65852,7 +65852,7 @@ DecodeJson(DecodeJsonCallbacks& callbacks,
               goto done;
             }
             depth++;
-            if (depth > WUFFS_JSON__DECODER_DEPTH_MAX_INCL) {
+            if (depth > (int32_t)WUFFS_JSON__DECODER_DEPTH_MAX_INCL) {
               ret_error_message =
                   "wuffs_aux::DecodeJson: internal error: bad depth";
               goto done;
